requirements.txt
Flask
Werkzeug
networkx
matplotlib
numpy
Flask-Mail
premailer
python-dotenv
APScheduler